requirements.txt
python-dotenv==0.19.2
schedule==1.2.2
selenium==4.23.0
